+++ server/src/domain/models.rs
@@ -5,8 +5,8 @@
 mod game;
 mod order;
 mod path;
-mod phase;
 mod phase_context;
+mod phases;
 mod player;
 mod power;
 mod province;
@@ -32,9 +32,9 @@ pub(crate) use order::OrderStatus;
 pub(crate) use order::RetreatOrder;
 pub(crate) use order::SupportOrder;
 pub(crate) use path::Path;
-pub(crate) use phase::Phase;
-pub(crate) use phase::PhaseKind;
 pub(crate) use phase_context::PhaseContext;
+pub(crate) use phases::Phase;
+pub(crate) use phases::PhaseKind;
 pub(crate) use player::Player;
 pub(crate) use power::Power;
 pub(crate) use province::Province;
