+++ server/src/domain/models/phases/phase.rs
@@ -1,34 +1,13 @@
 // ============================================================================
-// modules
-// ============================================================================
-
-mod methods;
-
-// ============================================================================
-// re-exports
-// ============================================================================
-
-pub(crate) use super::AdjustmentAdjudicator;
-pub(crate) use super::AdjustmentOrderHelper;
-pub(crate) use super::MainAdjudicator;
-pub(crate) use super::MainOrderHelper;
-pub(crate) use super::Order;
-pub(crate) use super::OrderKind;
-pub(crate) use super::PhaseContext;
-pub(crate) use super::Power;
-pub(crate) use super::Province;
-pub(crate) use super::RetreatAdjudicator;
-pub(crate) use super::RetreatOrderHelper;
-pub(crate) use super::Territory;
-pub(crate) use super::Unit;
-pub(crate) use super::UnitHelper;
-
-// ============================================================================
 // imports
 // ============================================================================
 
 use serde::Deserialize;
 use serde::Serialize;
+
+use super::Order;
+use super::Territory;
+use super::Unit;
 
 // ============================================================================
 // definitions
