+++ server/src/domain/models/phases/test_datc_6_f.rs
@@ -7,10 +7,10 @@
 //! [DATC_6F]: https://webdiplomacy.net/doc/DATC_v3_0.html#6.F
 
 use crate::domain::models::order::*;
-use crate::domain::models::phase::methods::*;
-use crate::domain::tests::a;
-use crate::domain::tests::f;
-use crate::domain::tests::p;
+use crate::domain::models::phases::method::*;
+use crate::domain::test::a;
+use crate::domain::test::f;
+use crate::domain::test::p;
 
 /// 6.F.1. TEST CASE, NO CONVOY IN COASTAL AREAS
 /// A fleet in a coastal area may not convoy.
