+++ server/src/domain/models/phases/test_datc_6_g.rs
@@ -7,10 +7,10 @@
 //! [DATC_6G]: https://webdiplomacy.net/doc/DATC_v3_0.html#6.G
 
 use crate::domain::models::order::*;
-use crate::domain::models::phase::methods::*;
-use crate::domain::tests::a;
-use crate::domain::tests::f;
-use crate::domain::tests::p;
+use crate::domain::models::phases::method::*;
+use crate::domain::test::a;
+use crate::domain::test::f;
+use crate::domain::test::p;
 
 /// 6.G.1. TEST CASE, TWO UNITS CAN SWAP PROVINCES BY CONVOY
 /// The only way to swap two units, is by convoy.
